# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Book sales platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/login:
    post:
      operationId: agentLogin
      tags: [agent]
      summary: Agent login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentLoginResult"
        "401":
          description: Invalid credentials
  /agent/me:
    get:
      operationId: getAgentMe
      tags: [agent]
      summary: Get current agent profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "401":
          description: Unauthorized
  /agent/orders:
    get:
      operationId: listAgentOrders
      tags: [agent]
      summary: List orders created by this agent
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Agent orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentOrdersResult"
        "401":
          description: Unauthorized
    post:
      operationId: createAgentOrder
      tags: [agent]
      summary: Create an order on behalf of a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentOrderInput"
      responses:
        "201":
          description: Order created with payment link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentOrderResult"
        "400":
          description: Validation error
        "409":
          description: Insufficient stock
  /agent/stats:
    get:
      operationId: getAgentStats
      tags: [agent]
      summary: Get agent sales summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStats"
        "401":
          description: Unauthorized
  /pay/{token}:
    get:
      operationId: getOrderByPaymentToken
      tags: [payments]
      summary: Get order details via payment link token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Payment link not found
...
  /courier/login:
    post:
      operationId: courierLogin
      tags: [courier]
      summary: Courier login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierLoginResult"
        "401":
          description: Invalid credentials
  /courier/me:
    get:
      operationId: getCourierMe
      tags: [courier]
      summary: Get current courier profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Courier profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierProfile"
        "401":
          description: Unauthorized
  /courier/orders:
    get:
      operationId: listCourierOrders
      tags: [courier]
      summary: List orders assigned to this courier
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Assigned orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrdersResult"
        "401":
          description: Unauthorized
  /courier/orders/{id}/status:
    patch:
      operationId: updateCourierOrderStatus
      tags: [courier]
      summary: Update delivery status for an assigned order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrderResult"
        "403":
          description: Order not assigned to this courier
        "404":
          description: Not found
  /courier/stats:
    get:
      operationId: getCourierStats
      tags: [courier]
      summary: Get courier delivery stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Courier stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierStats"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      required: [id, fullName, phone, email, deliveryAddress, city, productType, quantity, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        deliveryAddress:
          type: string
        city:
          type: string
        productType:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, awaiting_delivery, picked_up, in_transit, delivered, failed_delivery, cancelled]
        courierId:
          type: ["integer", "null"]
        trackingNotes:
          type: ["string", "null"]
        deliveryPaymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        payment:
          $ref: "#/components/schemas/Payment"
...
    Payment:
      type: object
      required: [id, orderId, method, amount, status, reference, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, successful, failed]
        reference:
          type: string
        phoneNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        gatewayTransactionId:
          type: ["string", "null"]
        gatewayStatus:
          type: ["string", "null"]
        gatewayCheckoutUrl:
          type: ["string", "null"]
        gatewayProviderReference:
          type: ["string", "null"]
        gatewayIdempotencyKey:
          type: ["string", "null"]
        gatewayLastPayload:
          type: ["object", "null"]
          additionalProperties: true
        lastCheckedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    AgentProfile:
      type: object
      required: [id, name, phone, email, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        active:
          type: boolean
        totalOrders:
          type: integer
        createdAt:
          type: string
...
    AgentCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AgentLoginResult:
      type: object
      required: [token, agent]
      properties:
        token:
          type: string
        agent:
          $ref: "#/components/schemas/AgentProfile"
    AgentOrderInput:
      type: object
      required: [fullName, phone, email, deliveryAddress, city, productType, quantity]
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        deliveryAddress:
          type: string
          minLength: 5
        city:
          type: string
          minLength: 2
        productType:
          type: string
          enum: [paperback, hardcover]
        quantity:
          type: integer
          minimum: 1
        notes:
          type: ["string", "null"]
        overrideTier:
          type: string
          enum: [standard, bulk, institutional]
    AgentOrderResult:
      type: object
      required: [id, fullName, phone, email, deliveryAddress, city, productType, quantity, totalAmount, status, pricingTier, paymentLink, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        deliveryAddress:
          type: string
        city:
          type: string
        productType:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        pricingTier:
          type: string
        paymentLink:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AgentOrdersResult:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/AgentOrderResult"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AgentStats:
      type: object
      required: [totalOrders, totalRevenue, pendingPayments, confirmedOrders]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingPayments:
          type: integer
        confirmedOrders:
          type: integer
...
    CourierProfile:
      type: object
      required: [id, name, phone, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        vehicleInfo:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    CourierCredentials:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    CourierLoginResult:
      type: object
      required: [token, courier]
      properties:
        token:
          type: string
        courier:
          $ref: "#/components/schemas/CourierProfile"
...
    CourierOrderResult:
      type: object
      required: [id, fullName, phone, deliveryAddress, city, productType, quantity, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        deliveryAddress:
          type: string
        city:
          type: string
        productType:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        trackingNotes:
          type: ["string", "null"]
        deliveryPaymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CourierOrdersResult:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/CourierOrderResult"
        total:
          type: integer
    CourierStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [picked_up, in_transit, delivered, failed_delivery]
        trackingNotes:
          type: ["string", "null"]
    CourierStats:
      type: object
      required: [assignedOrders, deliveredOrders, inTransitOrders, failedOrders]
      properties:
        assignedOrders:
          type: integer
        deliveredOrders:
          type: integer
        inTransitOrders:
          type: integer
        failedOrders:
          type: integer