# lib/api-zod/src/generated/types/agentProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book sales platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentProfile {
  id: number;
  name: string;
  phone: string;
  email: string;
  active: boolean;
  totalOrders?: number;
  createdAt: string;
}