# lib/api-zod/src/generated/types/courierCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book sales platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CourierCredentials {
  phone: string;
  password: string;
}