# lib/api-zod/src/generated/types/courierOrdersResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book sales platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CourierOrderResult } from './courierOrderResult';

export interface CourierOrdersResult {
  orders: CourierOrderResult[];
  total: number;
}