# lib/api-zod/src/generated/types/courierProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book sales platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CourierProfile {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  vehicleInfo?: string | null;
  active: boolean;
  createdAt: string;
}