# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book sales platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';
import type { Payment } from './payment';

export interface Order {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  deliveryAddress: string;
  city: string;
  productType: string;
  quantity: number;
  totalAmount: number;
  status: OrderStatus;
  /** @nullable */
  courierId?: number | null;
  /** @nullable */
  trackingNotes?: string | null;
  /** @nullable */
  deliveryPaymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  payment?: Payment;
}